model/requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
biopython
joblib
shap
catboost
torch